app/lightrag/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-multipart==0.0.20
httpx==0.28.1
pydantic==2.11.1
pydantic-settings==2.8.1
python-docx==1.1.2
pypdf==5.4.0
lightrag-hku==1.4.10
